Deprecation warning for ImmediateAsyncDispatcher → ImmediateDispatcher
Introduces RubyEventStore::ImmediateDispatcher as the new name.
Keeps ImmediateAsyncDispatcher as a deprecated subclass that warns on instantiation.

The rename clarifies intent: the dispatcher schedules work immediately,
it does not imply the work itself is async.

Related to: https://github.com/RailsEventStore/rails_event_store/pull/1561/commits/756e02184c2a2027cb0293ea305afaa9bf7bc7bf
