+++ ruby_event_store/lib/ruby_event_store.rb
@@ -42,6 +42,7 @@ require_relative "ruby_event_store/serializers/yaml"
 require_relative "ruby_event_store/batch_enumerator"
 require_relative "ruby_event_store/correlated_commands"
 require_relative "ruby_event_store/link_by_metadata"
+require_relative "ruby_event_store/immediate_dispatcher"
 require_relative "ruby_event_store/immediate_async_dispatcher"
 require_relative "ruby_event_store/composed_dispatcher"
 require_relative "ruby_event_store/composed_broker"
