+++ ruby_event_store/lib/ruby_event_store/immediate_dispatcher.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module RubyEventStore
+  class ImmediateDispatcher
+    def initialize(scheduler:)
+      @scheduler = scheduler
+    end
+
+    def call(subscriber, _, record)
+      @scheduler.call(subscriber, record)
+    end
+
+    def verify(subscriber)
+      @scheduler.verify(subscriber)
+    end
+  end
+end
