+++ ruby_event_store/spec/immediate_dispatcher_spec.rb
@@ -5,7 +5,7 @@ require "ruby_event_store/spec/dispatcher_lint"
 require "ruby_event_store/spec/scheduler_lint"
 
 module RubyEventStore
-  ::RSpec.describe ImmediateAsyncDispatcher do
+  ::RSpec.describe ImmediateDispatcher do
     class MyCustomScheduler
       def call(klass, record)
         klass.perform_async(record)
@@ -16,11 +16,7 @@ module RubyEventStore
       end
     end
 
-    it "warns when instantiated" do
-      expect { ImmediateAsyncDispatcher.new(scheduler: MyCustomScheduler.new) }.to output(/RubyEventStore::ImmediateAsyncDispatcher.*deprecated.*ImmediateDispatcher/m).to_stderr
-    end
-
-    it_behaves_like "dispatcher", ImmediateAsyncDispatcher.new(scheduler: MyCustomScheduler.new)
+    it_behaves_like "dispatcher", ImmediateDispatcher.new(scheduler: MyCustomScheduler.new)
     it_behaves_like "scheduler", MyCustomScheduler.new
 
     let(:event) { instance_double(Event) }
@@ -29,7 +25,7 @@ module RubyEventStore
 
     describe "#call" do
       specify do
-        dispatcher = ImmediateAsyncDispatcher.new(scheduler: scheduler)
+        dispatcher = ImmediateDispatcher.new(scheduler: scheduler)
 
         handler = spy
         dispatcher.call(handler, event, record)
@@ -40,14 +36,14 @@ module RubyEventStore
 
     describe "#verify" do
       specify do
-        dispatcher = ImmediateAsyncDispatcher.new(scheduler: scheduler)
+        dispatcher = ImmediateDispatcher.new(scheduler: scheduler)
 
         handler = double(perform_async: true)
         expect(dispatcher.verify(handler)).to be(true)
       end
 
       specify do
-        dispatcher = ImmediateAsyncDispatcher.new(scheduler: scheduler)
+        dispatcher = ImmediateDispatcher.new(scheduler: scheduler)
 
         handler = double
         expect(dispatcher.verify(handler)).to be(false)
